init ai sdk
