+++ package.json
@@ -37,9 +37,15 @@
   },
   "license": "MIT",
   "dependencies": {
+    "@ai-sdk/anthropic": "^2.0.17",
+    "@ai-sdk/google": "^2.0.14",
+    "@ai-sdk/mistral": "^2.0.14",
+    "@ai-sdk/openai": "^2.0.32",
+    "@ai-sdk/xai": "^2.0.20",
     "@effect/platform": "^0.90.9",
     "@effect/platform-node": "^0.96.1",
     "@effect/schema": "^0.75.5",
+    "ai": "^5.0.47",
     "chalk": "^5.6.2",
     "commander": "^14.0.1",
     "dotenv": "^17.2.2",
@@ -47,11 +53,11 @@
     "googleapis": "^159.0.0",
     "inquirer": "^12.9.4",
     "linkup-sdk": "^2.3.0",
-    "litellm": "^0.12.0",
     "marked": "^16.3.0",
     "marked-terminal": "^7.3.0",
     "open": "^10.2.0",
-    "short-uuid": "^5.0.1"
+    "short-uuid": "^5.0.1",
+    "zod": "^4.1.9"
   },
   "devDependencies": {
     "@types/bun": "^1.2.22",
